backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
anthropic==0.18.1
PyPDF2==3.0.1
pillow==10.2.0
pytesseract==0.3.10
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
google-auth==2.49.2
PyJWT==2.12.1
slowapi==0.1.9
